Move span tests into their own module.
